@@ -1,21 +1,21 @@
 name: dough
 description: This library provides widgets that can be used to create a smooshy user interface.
-version: 1.0.4
+version: 2.0.0
 repository: https://github.com/josiahsrc/dough/tree/master/packages/dough
 homepage: https://josiahsaunders.com/
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.2.4 <4.0.0"
+  flutter: ">=1.17.0"
 
 dependencies:
   equatable: ^2.0.5
   flutter:
     sdk: flutter
   provider: ^6.1.1
-  sensors: ^2.0.3
   vector_math: ^2.1.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  very_good_analysis: ^3.0.0
+  flutter_lints: ^2.0.0
