@@ -2,8 +2,5 @@ export 'src/dough/dough.dart';
 export 'src/dough/dough_controller.dart';
 export 'src/dough/dough_recipe.dart';
 export 'src/dough/dough_transformer.dart';
-export 'src/dough/draggable.dart';
-export 'src/dough/draggable_recipe.dart';
-export 'src/dough/gyro.dart';
-export 'src/dough/gyro_recipe.dart';
-export 'src/dough/pressable.dart';
+export 'src/dough/pressable_dough.dart';
+export 'src/dough/draggable_dough.dart';
